requirements.txt
Django==2.1.3
django-bootstrap-form==3.4
djangorestframework==3.9.0
html5lib==1.0.1
Pillow==5.3.0
PyPDF2==1.26.0
pytz==2018.7
reportlab==3.5.12
six==1.11.0
webencodings==0.5.1
xhtml2pdf==0.2.3
xlwt==1.3.0
